requirements.txt
psycopg2
psycopg2-binary
python-dotenv
fastapi[standard]
unicorn
email-validator